refact: create controlled textfield component, create login form component
